move other files than main in main package to their own packages
