+++ crates/gosub_styling/tools/generate_definitions/utils/stringmaybearray.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 import "encoding/json"
