+++ crates/gosub_styling/tools/generate_definitions/utils/types.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 type Data struct {
 	Properties []Properties `json:"properties"`
